+++ internal/musig2v040/README.md
@@ -1,0 +1,10 @@
+# MuSig2 v0.4.0
+
+This package contains an exact copy of the MuSig2 code as found in
+`github.com/btcsuite/btcec/v2/schnorr/musig2` at the tag `btcec/v2.2.2`.
+
+This corresponds to the [MuSig2 BIP specification version of
+`v0.4.0`](https://github.com/jonasnick/bips/blob/musig2/bip-musig2.mediawiki).
+
+We only keep this code here to allow implementing a backward compatible,
+versioned MuSig2 RPC.
