+++ internal/musig2v040/musig2_test.go
@@ -1,6 +1,6 @@
 // Copyright 2013-2022 The btcsuite developers
 
-package musig2
+package musig2v040
 
 import (
 	"bytes"
